+++ tests/xmpbox/type/test_abstract_field.py
@@ -27,6 +27,17 @@ def test_attribute_setters() -> None:
     assert attr.get_value() == "value2"
 
 
+def test_attribute_pdfbox_camelcase_aliases() -> None:
+    attr = Attribute("ns", "name", "value")
+    attr.setNsURI("ns2")
+    attr.setName("name2")
+    attr.setValue("value2")
+
+    assert attr.getNamespace() == "ns2"
+    assert attr.getName() == "name2"
+    assert attr.getValue() == "value2"
+
+
 def test_attribute_equality_and_repr() -> None:
     a = Attribute("ns", "name", "value")
     b = Attribute("ns", "name", "value")
@@ -43,6 +54,13 @@ def test_field_carries_property_name(metadata: XMPMetadata) -> None:
     assert field.get_property_name() == "renamed"
 
 
+def test_field_pdfbox_camelcase_property_aliases(metadata: XMPMetadata) -> None:
+    field = TextType(metadata, "ns", "p", "name", "v")
+    field.setPropertyName("renamed")
+
+    assert field.getPropertyName() == "renamed"
+
+
 def test_field_attribute_round_trip(metadata: XMPMetadata) -> None:
     field = TextType(metadata, "ns", "p", "name", "v")
     attr = Attribute("http://x/", "lang", "en-US")
@@ -53,6 +71,18 @@ def test_field_attribute_round_trip(metadata: XMPMetadata) -> None:
     field.remove_attribute("lang")
     assert field.contains_attribute("lang") is False
     assert field.get_attribute("lang") is None
+
+
+def test_field_pdfbox_camelcase_attribute_aliases(metadata: XMPMetadata) -> None:
+    field = TextType(metadata, "ns", "p", "name", "v")
+    attr = Attribute("http://x/", "lang", "en-US")
+    field.setAttribute(attr)
+
+    assert field.containsAttribute("lang") is True
+    assert field.getAttribute("lang") is attr
+    assert attr in field.getAllAttributes()
+    field.removeAttribute("lang")
+    assert field.containsAttribute("lang") is False
 
 
 def test_field_attribute_replaces_by_name(metadata: XMPMetadata) -> None:
@@ -66,3 +96,4 @@ def test_field_attribute_replaces_by_name(metadata: XMPMetadata) -> None:
 def test_field_metadata_back_reference(metadata: XMPMetadata) -> None:
     field = TextType(metadata, "ns", "p", "name", "v")
     assert field.get_metadata() is metadata
+    assert field.getMetadata() is metadata